+++ internal/cli/output_contract.go
@@ -32,7 +32,9 @@ conf apply
 conf attachment get
 conf attachment list
 conf blog create
+conf comment add
 conf comment list
+conf comment preview
 conf comment thread
 conf diff
 conf edit